# ai-service-openapi.yaml
openapi: 3.0.3
info:
  title: Knova AI Service API
  version: 1.0.0-draft
  description: |
    Contract for the standalone Python AI service.
    Covers compatibility endpoints, native AI tasks, editable projects,
    jobs, artifacts, templates, and media rendering workflows.
servers:
  - url: https://ai-service.internal
    description: Internal AI service
security:
  - bearerAuth: []
...
paths:
...
  /v1/providers:
    get:
      tags: [Catalog]
      summary: List configured providers
      operationId: listProviders
      responses:
        '200':
          description: Providers catalog
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderListResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /v1/models:
    get:
      tags: [Catalog, Compatibility]
      summary: List available model aliases and capabilities
      operationId: listModels
      responses:
        '200':
          description: Models catalog
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelListResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /v1/chat/completions:
    post:
      tags: [Compatibility]
      summary: OpenAI-style chat completions endpoint
      operationId: createChatCompletion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
      responses:
        '200':
          description: Completion generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
  /v1/embeddings:
    post:
      tags: [Compatibility]
      summary: OpenAI-style embeddings endpoint
      operationId: createEmbeddings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmbeddingsRequest'
      responses:
        '200':
          description: Embeddings created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmbeddingsResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
  /v1/audio/transcriptions:
    post:
      tags: [Compatibility]
      summary: Transcribe audio or extract speech from video
      operationId: createTranscription
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TranscriptionMultipartRequest'
      responses:
        '200':
          description: Transcription result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranscriptionResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
  /v1/audio/speech:
    post:
      tags: [Compatibility]
      summary: Synthesize speech and return artifact references
      operationId: createSpeechSynthesis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpeechSynthesisRequest'
      responses:
        '200':
          description: Speech generated synchronously
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpeechSynthesisResponse'
        '202':
          description: Speech generation accepted as async job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAcceptedResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
  /v1/tasks/chat-rag:
    post:
      tags: [Tasks]
      summary: Generate grounded chat answer from history and source chunks
      operationId: runChatRagTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRagRequest'
      responses:
        '200':
          description: Chat answer generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatRagResponse'
        '202':
          description: Task accepted as async job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAcceptedResponse'
  /v1/tasks/compare:
    post:
      tags: [Tasks]
      summary: Compare two or more documents
      operationId: runCompareTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompareRequest'
      responses:
        '200':
          description: Comparison result generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompareResponse'
        '202':
          description: Task accepted as async job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAcceptedResponse'
  /v1/tasks/mindmap:
    post:
      tags: [Tasks]
      summary: Generate editable mindmap graph
      operationId: runMindmapTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MindmapRequest'
      responses:
        '200':
          description: Mindmap generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MindmapResponse'
        '202':
          description: Task accepted as async job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAcceptedResponse'
  /v1/tasks/quiz:
    post:
      tags: [Tasks]
      summary: Generate quiz without fixed question limits
      operationId: runQuizTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizRequest'
      responses:
        '200':
          description: Quiz generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizResponse'
        '202':
          description: Task accepted as async job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAcceptedResponse'
  /v1/tasks/table:
    post:
      tags: [Tasks]
      summary: Extract structured table from arbitrary sources
      operationId: runTableTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TableRequest'
      responses:
        '200':
          description: Table generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableResponse'
        '202':
          description: Task accepted as async job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAcceptedResponse'
  /v1/tasks/infographic:
    post:
      tags: [Tasks]
      summary: Generate infographic specification and optional renders
      operationId: runInfographicTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InfographicRequest'
      responses:
        '200':
          description: Infographic generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfographicResponse'
        '202':
          description: Task accepted as async job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAcceptedResponse'
  /v1/tasks/report:
    post:
      tags: [Tasks]
      summary: Generate report from template and sources
      operationId: runReportTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportRequest'
      responses:
        '200':
          description: Report generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '202':
          description: Task accepted as async job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAcceptedResponse'
  /v1/tasks/presentation:
    post:
      tags: [Tasks]
      summary: Create editable presentation project
      operationId: runPresentationTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresentationTaskRequest'
      responses:
        '200':
          description: Presentation project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresentationTaskResponse'
        '202':
          description: Task accepted as async job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAcceptedResponse'
  /v1/tasks/podcast:
    post:
      tags: [Tasks]
      summary: Create editable podcast project with TTS pipeline
      operationId: runPodcastTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PodcastTaskRequest'
      responses:
        '200':
          description: Podcast project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PodcastTaskResponse'
        '202':
          description: Task accepted as async job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAcceptedResponse'
  /v1/tasks/video:
    post:
      tags: [Tasks]
      summary: Create editable video project or rendering job
      operationId: runVideoTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoTaskRequest'
      responses:
        '200':
          description: Video project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoTaskResponse'
        '202':
          description: Task accepted as async job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAcceptedResponse'
  /v1/projects:
    get:
      tags: [Projects]
      summary: List AI projects
      operationId: listProjects
      parameters:
        - in: query
          name: project_type
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: cursor
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            default: 50
      responses:
        '200':
          description: Project list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
    post:
      tags: [Projects]
      summary: Create empty or seeded AI project
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreateRequest'
      responses:
        '200':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectEnvelope'
        '400':
          $ref: '#/components/responses/ErrorResponse'
...
  /v1/projects/{project_id}/export:
    post:
      tags: [Projects]
      summary: Export project into one or more artifacts
      operationId: exportProject
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
      responses:
        '200':
          description: Export completed synchronously
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportResponse'
        '202':
          description: Export accepted as async job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAcceptedResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    ProviderListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProviderDescriptor'
...
    ModelListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ModelDescriptor'
...
    EmbeddingsRequest:
      type: object
      required: [input]
      properties:
        model:
          type: string
          default: embedding-default
        provider:
          type: string
          default: auto
        input:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        dimensions:
          type: integer
          minimum: 1
        task_type:
          type: string
          enum:
            - retrieval_query
            - retrieval_document
            - semantic_similarity
            - classification
            - clustering
        encoding_format:
          type: string
          enum: [float]
          default: float
...